before multiple-inheritence

+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/interpreter/logo_lang/interpreter-comp/kmLogo/interpreter/kmLogo/impl/ClearImpl.java
@@ -18,7 +18,7 @@ public class ClearImpl extends PrimitiveImpl implements Clear {
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = 0.0;
+		result = (double) (0.0) ;
 		return result;
 	}
 }
