+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/interpreter/logo_lang/interpreter-comp/kmLogo/interpreter/kmLogo/impl/EqualsImpl.java
@@ -20,11 +20,11 @@ public class EqualsImpl extends BinaryExpImpl implements Equals {
 
 	public double eval(Turtle turtle) {
 		double result;
-		if (EqualService.equals((((Expression) this.lhs).eval((Turtle) (turtle))), (((Expression) this.rhs).eval((Turtle) (turtle))))) {
-			result = 1.0;
+		if (EqualService.equals((((Expression) (this.lhs)).eval((Turtle) (turtle))), (((Expression) (this.rhs)).eval((Turtle) (turtle))))) {
+			result = (double) (1.0) ;
 		}
 		else {
-			result = 0.0;
+			result = (double) (0.0) ;
 		}
 		return result;
 	}
