+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/interpreter/logo_lang/interpreter-comp/kmLogo/interpreter/kmLogo/impl/ExpressionImpl.java
@@ -18,7 +18,7 @@ public abstract class ExpressionImpl extends InstructionImpl implements Expressi
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = 0.0;
+		result = (double) (0.0) ;
 		return result;
 	}
 }
