+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/interpreter/logo_lang/interpreter-comp/kmLogo/interpreter/kmLogo/impl/GreaterImpl.java
@@ -19,11 +19,11 @@ public class GreaterImpl extends BinaryExpImpl implements Greater {
 
 	public double eval(Turtle turtle) {
 		double result;
-		if ((((Expression) this.lhs).eval((Turtle) (turtle))) > (((Expression) this.rhs).eval((Turtle) (turtle)))) {
-			result = 1.0;
+		if ((((Expression) (this.lhs)).eval((Turtle) (turtle))) > (((Expression) (this.rhs)).eval((Turtle) (turtle)))) {
+			result = (double) (1.0) ;
 		}
 		else {
-			result = 0.0;
+			result = (double) (0.0) ;
 		}
 		return result;
 	}
