+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/interpreter/logo_lang/interpreter-comp/kmLogo/interpreter/kmLogo/impl/InstructionImpl.java
@@ -19,7 +19,7 @@ public abstract class InstructionImpl extends MinimalEObjectImpl.Container imple
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = 0.0;
+		result = (double) (0.0) ;
 		return result;
 	}
 }
