+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/interpreter/logo_lang/interpreter-comp/kmLogo/interpreter/kmLogo/impl/MultImpl.java
@@ -19,7 +19,7 @@ public class MultImpl extends BinaryExpImpl implements Mult {
 
 	public double eval(Turtle turtle) {
 		double result;
-		result = (((Expression) this.lhs).eval((Turtle) (turtle))) * (((Expression) this.rhs).eval((Turtle) (turtle)));
+		result = (double) ((((Expression) (this.lhs)).eval((Turtle) (turtle))) * (((Expression) (this.rhs)).eval((Turtle) (turtle)))) ;
 		return result;
 	}
 }
