+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/interpreter/logo_lang/interpreter-comp/kmLogo/interpreter/kmLogo/impl/PenUpImpl.java
@@ -19,7 +19,7 @@ public class PenUpImpl extends PrimitiveImpl implements PenUp {
 	public double eval(Turtle turtle) {
 		double result;
 		turtle.setPenUp(true);
-		result = 0.0;
+		result = (double) (0.0) ;
 		return result;
 	}
 }
